testing/jamfpy/requirements.txt
python-dotenv
jamfpy @ git+https://github.com/thejoeker12/jamfpy-python-sdk-jamfpro@dev